# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_kishert-lab_taxi-platform_internal_dto.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOi...
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthConfirmCodeRequest:
    properties:
      code:
        example: "1234"
        maxLength: 6
        minLength: 4
        type: string
      name:
        example: Сергей
        maxLength: 255
        type: string
      phone:
        example: "+79124966126"
        type: string
    required:
    - code
    - phone
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOi...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOi...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeRequest:
    properties:
      phone:
        example: "+79124966126"
        type: string
    required:
    - phone
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeResponse:
    properties:
      message:
        example: Код подтверждения отправлен
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOi...
        type: string
      expires_in:
        example: 900
        type: integer
      passenger:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerMeResponse'
      refresh_token:
        example: eyJhbGciOi...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerCreateOrderRequest:
    properties:
      car_class_id:
        example: 33333333-3333-3333-3333-333333333333
        type: string
      city_id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
      comment:
        example: Entrance 2
        type: string
      destination_address:
        example: Mira 10
        type: string
      destination_location:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.CoordinatesRequest'
      passenger_location_sharing_enabled:
        example: false
        type: boolean
      passenger_phone:
        example: "+79990000000"
        type: string
      payment_type:
        allOf:
        - $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_domain.PaymentMethod'
        enum:
        - cash
        - card
        - corporate
        example: cash
      pickup_address:
        example: Lenina 1
        type: string
      pickup_comment:
        example: Вход со двора
        type: string
      pickup_entrance:
        example: "2"
        type: string
      pickup_location:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.CoordinatesRequest'
      tariff_id:
        example: 22222222-2222-2222-2222-222222222222
        type: string
    required:
    - destination_address
    - destination_location
    - payment_type
    - pickup_address
    - pickup_location
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerMePatchRequest:
    properties:
      avatar_url:
        example: https://cdn.example.com/passengers/111/avatar.jpg
        type: string
      email:
        example: test@example.com
        type: string
      name:
        example: Сергей
        maxLength: 255
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenRequest:
    properties:
      device_id:
        example: pixel-8-pro
        type: string
      platform:
        enum:
        - android
        - ios
        - web
        example: android
        type: string
      token:
        example: fcm_device_token
        type: string
    required:
    - platform
    - token
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOi...
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_maps.ReverseResponse:
    properties:
      address:
        type: string
      address_location:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
      found:
        type: boolean
      point:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
      source:
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_maps.RouteRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_geocoder_domain.Coordinates'
        type: array
    required:
    - points
    type: object
...
  github_com_kishert-lab_taxi-platform_pkg_response.Error:
    properties:
      error:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  github_com_kishert-lab_taxi-platform_pkg_response.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.ErrorCode'
        example: ORDER_NOT_FOUND
      details:
        additionalProperties: {}
        type: object
      message:
        example: Order not found
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_pkg_response.Meta:
    properties:
      request_id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
    type: object
...
  internal_transport_http_handler.MapReverseResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.ReverseResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.MapRouteResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_routing.Route'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.PassengerAddressSearchResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/internal_transport_http_handler.PassengerAddressSearchResultResponse'
        type: array
    type: object
...
  internal_transport_http_handler.PassengerAddressSearchSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_handler.PassengerAddressSearchResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerAuthRefreshSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRefreshResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerAuthRequestCodeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerAuthTokenSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthTokenResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerCarClassesSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerCarClassesResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerLogoutSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_handler.passengerLogoutResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  internal_transport_http_handler.PassengerMeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerMeResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.PassengerOrderSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerOrderResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.PassengerPushTokenSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.passengerLogoutResponse:
    properties:
      logged_out:
        example: true
        type: boolean
    type: object
info:
  contact:
    email: support@example.com
    name: Taxi Platform API Support
  description: Production-grade taxi backend API for small cities and rural areas.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taxi Platform API
  version: 0.1.0
paths:
...
  /passenger/address/reverse:
    get:
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapReverseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Reverse geocode without changing the selected pickup point
      tags:
      - maps
  /passenger/address/search:
    get:
      description: Searches address suggestions for passenger mobile order form.
      parameters:
      - description: Bearer passenger access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Address query
        in: query
        name: q
        required: true
        type: string
      - description: City UUID
        in: query
        name: city_id
        type: string
      - description: Focus latitude
        in: query
        name: lat
        type: number
      - description: Focus longitude
        in: query
        name: lon
        type: number
      - description: Result limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAddressSearchSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Search passenger addresses
      tags:
      - passenger
  /passenger/auth/confirm-code:
    post:
      consumes:
      - application/json
      description: Verifies the one-time code and returns passenger access and refresh
        tokens.
      parameters:
      - description: Passenger code confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthConfirmCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAuthTokenSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Confirm passenger authentication code
      tags:
      - passenger-auth
  /passenger/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Passenger refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerLogoutSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Logout passenger
      tags:
      - passenger-auth
  /passenger/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Passenger refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAuthRefreshSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Refresh passenger tokens
      tags:
      - passenger-auth
  /passenger/auth/request-code:
    post:
      consumes:
      - application/json
      description: Sends a one-time confirmation code to the passenger phone number.
      parameters:
      - description: Passenger code request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerAuthRequestCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerAuthRequestCodeSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      summary: Request passenger authentication code
      tags:
      - passenger-auth
  /passenger/car-classes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerCarClassesSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: List active passenger car classes
      tags:
      - passenger-orders
  /passenger/map/routes:
    post:
      consumes:
      - application/json
      description: Points use latitude/longitude fields. GeoJSON geometry uses [longitude,latitude].
        Original points are preserved; snapped points are separate. No live traffic.
        Does not grant access to order or driver coordinates; obtain those through
        existing authorized order and WebSocket APIs.
      parameters:
      - description: Exactly two points; order model has no intermediate stops
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_maps.RouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.MapRouteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Calculate a road route between two supplied points
      tags:
      - maps
  /passenger/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerMeSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Get current passenger profile
      tags:
      - passenger
    patch:
      consumes:
      - application/json
      parameters:
      - description: Passenger profile patch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerMePatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerMeSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Update current passenger profile
      tags:
      - passenger
  /passenger/orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerCreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerOrderSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Create passenger order
      tags:
      - passenger-orders
...
  /passenger/push-tokens:
    post:
      consumes:
      - application/json
      description: Registers or updates the push token for the authenticated passenger
        device.
      parameters:
      - description: Passenger push token registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerPushTokenSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Register passenger push token
      tags:
      - passenger
  /passenger/push/token:
    post:
      consumes:
      - application/json
      description: Registers or updates the push token for the authenticated passenger
        device.
      parameters:
      - description: Passenger push token registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerPushTokenSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Register passenger push token
      tags:
      - passenger
...
  /ws:
    get:
      description: Mobile realtime endpoint. JWT can be passed in Authorization header
        or token query parameter. After reconnect the server emits sync.required and
        the client must call the current order REST endpoint.
      parameters:
      - description: JWT token fallback for mobile clients
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Connect mobile WebSocket
      tags:
      - websocket
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer ` prefix, for example: Bearer eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"